refactor(protocol): reorganize protocol system into modular components
- Split monolithic protocol.rs into dedicated modules
- Create layered architecture with base.rs, global.rs, local.rs, transforms.rs, utils.rs
- Fix all code style issues with cargo fmt and clippy
- Update documentation to reflect new structure
- Add markdown linting task to TASKS.md for future improvement
