+++ src/protocol/base.rs
@@ -10,9 +10,69 @@
 //! These types form the foundation for the type-level programming patterns
 //! used throughout the protocol system.
 
-use crate::sealed;
-use crate::types;
 use core::marker::PhantomData;
 
-// Contents will be migrated from protocol.rs
-// during the code migration phase
+/// # Type-Level Map/Fold Pattern
+///
+/// Many traits in this crate use a type-level map/fold pattern to recursively
+/// process type-level lists or protocol structures. This is a common idiom for
+/// building, transforming, or checking protocol types at compile time.
+///
+/// ## How it works
+/// - The trait is implemented for the base case (usually `Nil`), which provides
+///   the default or terminal value.
+/// - The trait is then implemented recursively for `Cons<H, T>` (or similar),
+///   where the head is processed and the result is combined with the recursive
+///   result for the tail.
+///
+///   Type-level empty list for n-ary combinators and role/label sets.
+///   Used as the base case for type-level lists.
+pub struct Nil;
+
+/// `Cons<H, T>` is a type-level list node with a head type `H` and a tail type `T`.
+///
+/// The tail must itself be a type-level list. The implementation typically recursively processes the head and
+/// continues the recursion on the tail.
+///
+/// # Type parameters
+///
+/// - `H`: The head type in this node
+/// - `T`: The tail type-level list
+pub struct Cons<H, T>(PhantomData<(H, T)>);
+
+/// Trait to check that all elements in a type-level list are unique.
+///
+/// Used for compile-time uniqueness assertions (e.g., for protocol labels).
+pub trait UniqueList {}
+
+impl UniqueList for Nil {}
+impl<H, T> UniqueList for Cons<H, T> where T: NotInList<H> + UniqueList {}
+
+/// Helper trait to check if a type is not in a type-level list.
+pub trait NotInList<X> {}
+
+impl<X> NotInList<X> for Nil {}
+impl<X, H, T> NotInList<X> for Cons<H, T>
+where
+    X: NotSame<H>,
+    T: NotInList<X>,
+{
+}
+
+/// Helper trait to check if two types are not the same.
+pub trait NotSame<T> {}
+
+impl<A, B> NotSame<B> for A where A: NotTypeEq<B> {}
+
+/// Base trait for type inequality.
+/// The default implementation works for all non-identical types.
+pub trait NotTypeEq<B> {}
+
+impl<A, B> NotTypeEq<B> for A {}
+// Overlap: no impl for A == A (this is intentional)
+
+/// Type-level trait to check if two types are the same.
+/// Implemented via a marker type Output that is True or False.
+pub trait TypeEq<B> {
+    type Output;
+}
