+++ src/protocol/mod.rs
@@ -1,33 +1,45 @@
-//! # Protocol Module
+//! # Protocol System
 //!
-//! This module defines the core protocol types and type-level programming patterns
-//! for session types in Besedarium. It provides the foundation for specifying,
-//! validating, and projecting communication protocols.
+//! This module provides a type-level approach to designing, verifying,
+//! and implementing communication protocols using session types.
 //!
-//! The module is split into several components based on abstraction layers:
+//! ## Module Structure
 //!
-//! * `base`: Foundational types like `Nil`, `Cons` and base traits
-//! * `global`: Global protocol types (`TSession`, `TEnd`, `TInteract`, etc.)
-//! * `local`: Local (endpoint) protocol types (`EpSend`, `EpRecv`, etc.)
-//! * `transforms`: Transformation logic including projection (`ProjectRole`, etc.)
-//! * `utils`: Helper traits and type-level operations
+//! - `base`: Foundational types and traits for type-level programming
+//! - `global`: Global protocol types representing multi-party choreography
+//! - `local`: Local protocol types representing endpoint behavior
+//! - `transforms`: Projection and other transformations between protocol representations
+//! - `utils`: Utility traits for protocol manipulation and checking
 //!
-//! ## Type-Level Programming
+//! ## Key Concepts
 //!
-//! Many traits in this module use a type-level map/fold pattern to recursively
-//! process type-level lists or protocol structures. This enables compile-time
-//! verification of protocol properties and transformations.
+//! - **Global Protocols**: Describe the overall choreography between participants
+//! - **Local Protocols**: Describe the behavior of a single participant
+//! - **Projection**: The process of deriving local protocols from global ones
+//! - **Type-Level Operations**: Compile-time reasoning about protocol properties
 
-// Declare submodules
-mod base;
-mod global;
-mod local;
-mod transforms;
-mod utils;
+// Re-export everything from the submodules
+pub mod base;
+pub mod global;
+pub mod local;
+pub mod transforms;
+pub mod utils;
 
-// Re-export everything from submodules
-pub use base::*;
-pub use global::*;
-pub use local::*;
-pub use transforms::*;
-pub use utils::*;
+// Re-export commonly used items at the protocol module level
+pub use self::base::{Cons, Nil, NotInList, NotSame, NotTypeEq, UniqueList};
+pub use self::global::{
+    AssertDisjoint, TChoice, TEnd, TInteract, TPar, TRec, TSession, ToTChoice, ToTPar,
+};
+pub use self::local::{
+    EpChoice, EpEnd, EpPar, EpRecv, EpSend, EpSession, EpSkip, GetEpSkipTypeMarker, IsEnd,
+    IsEpEndVariant, IsEpSkipTypeImpl, IsEpSkipVariant, IsSkip, Role, RoleEq, TBroker, TClient,
+    TServer, TWorker, Void,
+};
+pub use self::transforms::{
+    ComposeProjectedParBranches, ComposeProjectedParBranchesCase, ContainsRole, FilterSkips,
+    FilterSkipsCase, NotContainsRole, ProjectChoice, ProjectChoiceCase, ProjectInteract,
+    ProjectPar, ProjectParBranch, ProjectRole, TParContainsRoleImpl,
+};
+pub use self::utils::{
+    CheckNil, Concat, ConcatCons, Disjoint, DisjointCons, IsEmpty, IsNil, IsNotNil,
+};
