+++ work/TASKS.md
@@ -2,15 +2,18 @@
 
 ## Active Tasks
 
-- [ ] (wip) Refactor `protocol.rs` into smaller submodules
-  - [x] Create plan and structure for protocol module refactoring
-  - [x] Create GitHub issue documenting the refactoring need (Issue #11)
-  - [x] Create draft pull request for implementation (PR #12) 
-  - [x] Implement initial layer-based module structure
-  - [ ] Migrate code from protocol.rs to appropriate modules
-  - [ ] Complete comprehensive testing of the refactored structure
-  - [ ] Merge changes back to main branch
+- [ ] Fix markdown linting issues across documentation files
+  - [ ] Address line length issues (over 80 characters)
+  - [ ] Fix list formatting issues (missing blank lines)
+  - [ ] Resolve heading formatting issues
+  - [ ] Fix minor issues (trailing spaces, missing newlines)
 
 ## Completed Tasks
 
-- None yet
+- [x] Refactor `protocol.rs` into smaller submodules
+  - [x] Create plan and structure for protocol module refactoring
+  - [x] Create GitHub issue documenting the refactoring need (Issue #11)
+  - [x] Create draft pull request for implementation (PR #12)
+  - [x] Implement initial layer-based module structure
+  - [x] Migrate code from protocol.rs to appropriate modules
+  - [x] Complete comprehensive testing of the refactored structure
